# server/swagger/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "AnythingLLM Developer API",
    "description": "API endpoints that enable programmatic reading, writing, and updating of your AnythingLLM instance. UI supplied by Swagger.io."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/v1/embed/{uuid}": {
      "get": {
        "tags": [
          "Embed"
        ],
        "description": "Get a single embed by UUID",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "UUID of the embed"
          }
        ],
        "responses": {
          "200": {
            "description": "Embed configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbedConfig"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Embed not found"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/v1/embed/{embedUuid}/chats": {
      "get": {
        "tags": [
          "Embed"
        ],
        "description": "Get all chats for a specific embed",
        "parameters": [
          {
            "name": "embedUuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "UUID of the embed"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "chats": [
                      {
                        "id": 1,
                        "session_id": "session-uuid-1",
                        "prompt": "Hello",
                        "response": "Hi there!",
                        "createdAt": "2023-04-01T12:00:00Z"
                      },
                      {
                        "id": 2,
                        "session_id": "session-uuid-2",
                        "prompt": "How are you?",
                        "response": "I'm doing well, thank you!",
                        "createdAt": "2023-04-02T14:30:00Z"
                      }
                    ]
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidAPIKey"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidAPIKey"
                }
              }
            }
          },
          "404": {
            "description": "Embed not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/v1/embed/{embedUuid}/chats/{sessionUuid}": {
      "get": {
        "tags": [
          "Embed"
        ],
        "description": "Get chats for a specific embed and session",
        "parameters": [
          {
            "name": "embedUuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "UUID of the embed"
          },
          {
            "name": "sessionUuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "UUID of the session"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "chats": [
                      {
                        "id": 1,
                        "prompt": "Hello",
                        "response": "Hi there!",
                        "createdAt": "2023-04-01T12:00:00Z"
                      }
                    ]
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidAPIKey"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidAPIKey"
                }
              }
            }
          },
          "404": {
            "description": "Embed or session not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "InvalidAPIKey": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Invalid API Key"
          }
        },
        "xml": {
          "name": "InvalidAPIKey"
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}